Updated to 0.113u2 core
git-svn-id: svn://messdev.no-ip.org/mess@118 9849079f-6229-0410-a07f-84406ce305ba

+++ src/osd/windows/ui/properties.c
@@ -749,12 +749,12 @@ static char *GameInfoScreen(UINT nIndex)
 			sprintf(buf,"%d x %d (V) %f Hz",
 			drv.screen[0].defstate.visarea.max_y - drv.screen[0].defstate.visarea.min_y + 1,
 					drv.screen[0].defstate.visarea.max_x - drv.screen[0].defstate.visarea.min_x + 1,
-					drv.screen[0].defstate.refresh);
+					(float) drv.screen[0].defstate.refresh);
 		else
 			sprintf(buf,"%d x %d (H) %f Hz",
 					drv.screen[0].defstate.visarea.max_x - drv.screen[0].defstate.visarea.min_x + 1,
 					drv.screen[0].defstate.visarea.max_y - drv.screen[0].defstate.visarea.min_y + 1,
-					drv.screen[0].defstate.refresh);
+					(float) drv.screen[0].defstate.refresh);
 	}
 	return buf;
 }
